refactor: Update navigation links in nav.ts
This commit updates the navigation links in the nav.ts file. The "Home" link is changed to "Blog", the "Account" link is changed to "About", the "Settings" link is changed to "Help", and a new link "Careers" is added. This change improves the navigation menu by providing more relevant and descriptive links.

Note: This commit message has been condensed and formatted to adhere to the repository's commit message conventions.