chore: Update npm dependencies
This commit updates the npm dependencies to the latest stable versions. Specifically, it adds the following packages:
- @tanstack/react-query@5.51.23
- @tanstack/react-query-devtools@5.51.23
- @types/ua-parser-js@0.7.39
- ua-parser-js@1.0.38

These updates ensure that the project is using the most recent versions of these packages, which may include bug fixes, performance improvements, and new features.

Note: This commit message has been condensed and formatted to adhere to the repository's commit message conventions.