refactor: Update default country name fetching in SignUpForm
This commit refactors the SignUpForm component in the SignUpForm.tsx file. It updates the default country name fetching logic to use the new API endpoint '/api/get-country/name' and properly handle the response data. The previous implementation had a bug where it was not correctly setting the default country name. This change improves the functionality and reliability of the SignUpForm component.

Note: This commit message has been condensed and formatted to adhere to the repository's commit message conventions.