refactor: Update privateLinks in NavbarSingle component
This commit updates the privateLinks array in the NavbarSingle component to include a new link for support. The link is added to the list of privateLinks and is displayed in the navigation menu. The target attribute is set to "_blank" to open the support page in a new tab. This change enhances the functionality of the NavbarSingle component by providing a direct link to the support page for users.

Note: This commit message has been condensed and formatted to adhere to the repository's commit message conventions.