refactor: Update formatRelativeDate function and add formattedDateTime function
This commit refactors the formatRelativeDate function in the utils.ts file to include a return type annotation. It also adds the formattedDateTime function to the same file, which formats a given date as "MMMM d, yyyy 'at' h:mm a". These changes improve the code clarity and maintainability of the utils module.

Note: This commit message has been condensed and formatted to adhere to the repository's commit message conventions.