refactor: Update ReferralsDetails component to copy referral link with query parameter
This commit updates the ReferralsDetails component in the account/referrals directory. It modifies the onClick event handler of the copy button to copy the referral link with a query parameter. The query parameter includes the user's username or a default value if the username is not available. This change improves the functionality of the referral feature by providing a direct link to the referral page for users.

Note: This commit message has been condensed and formatted to adhere to the repository's commit message conventions.