refactor: Update Information component to use updateProfile function
This commit updates the Information component in the src/app/(user)/account/Information.tsx file to use the updateProfile function from the actions module. The updateProfile function is imported and called in the onSubmit function, replacing the previous axios.put request. This refactor improves the code organization and maintainability by separating the API call logic into a separate module.

Note: This commit message has been condensed and formatted to adhere to the repository's commit message conventions.